:recycle: refactor(seller): test code 수정 및 상품 등록 시 null 값을 허용하지 않게 수정 상품 수량 수정 -> 상품 수정으로 합쳤습니다.

+++ src/main/java/com/kuku9/goods/domain/seller/service/SellerServiceImpl.java
@@ -8,7 +8,6 @@ import com.kuku9.goods.domain.product.repository.ProductRepository;
 import com.kuku9.goods.domain.search.document.ProductDocument;
 import com.kuku9.goods.domain.search.dto.ProductSearchResponse;
 import com.kuku9.goods.domain.search.repository.ProductSearchRepository;
-import com.kuku9.goods.domain.seller.dto.request.ProductQuantityRequest;
 import com.kuku9.goods.domain.seller.dto.request.ProductRegistRequest;
 import com.kuku9.goods.domain.seller.dto.request.ProductUpdateRequest;
 import com.kuku9.goods.domain.seller.dto.response.SellerCheckResponse;
@@ -43,7 +42,7 @@ public class SellerServiceImpl implements SellerService {
 
     @Override
     @Transactional
-    public Long createProduct(ProductRegistRequest requestDto, User user) {
+    public String createProduct(ProductRegistRequest requestDto, User user) {
         Seller seller = findSeller(user);
 
         Product product = new Product(requestDto, seller);
@@ -57,38 +56,26 @@ public class SellerServiceImpl implements SellerService {
 
         productSearchRepository.save(productDocument);
 
-        return saveProduct.getSeller().getId();
+        return saveProduct.getSeller().getDomainName();
     }
 
     @Override
     @Transactional
     @CacheEvict(value = "productCache", key = "#productId")
-    public Long updateProductStatus(Long productId, User user) {
+    public String updateProductStatus(Long productId, User user) {
         Seller seller = findSeller(user);
 
         Product product = findProduct(productId, seller);
 
         product.updateOrderStatus(product.getStatus());
 
-        return product.getSeller().getId();
-    }
-
-    @Override
-    @Transactional
-    public Long updateProductQuantity(Long productId, ProductQuantityRequest request, User user) {
-        Seller seller = findSeller(user);
-
-        Product product = findProduct(productId, seller);
-
-        product.updateQuantitySeller(request);
-
-        return product.getSeller().getId();
+        return product.getSeller().getDomainName();
     }
 
     @Override
     @Transactional
     @CacheEvict(value = "productCache", key = "#productId")
-    public Long updateProduct(
+    public String updateProduct(
         Long productId, ProductUpdateRequest requestDto, User user) {
         String productName, productDescription;
         int price, quantity;
@@ -97,27 +84,32 @@ public class SellerServiceImpl implements SellerService {
 
         Product product = findProduct(productId, seller);
 
-        productSearchRepository.deleteByProductId(product.getId());
+        if (requestDto != null) {
 
-        if (requestDto.getName().isEmpty()) {
-            productName = product.getName();
+            productSearchRepository.deleteByProductId(product.getId());
+
+            if (requestDto.getName().isEmpty()) {
+                productName = product.getName();
+            } else {
+                productName = requestDto.getName();
+            }
+            if (requestDto.getDescription().isEmpty()) {
+                productDescription = product.getDescription();
+            } else {
+                productDescription = requestDto.getDescription();
+            }
+            if (requestDto.getPrice() <= 0) {
+                price = product.getPrice();
+            } else {
+                price = requestDto.getPrice();
+            }
+            if (requestDto.getQuantity() <= 0) {
+                quantity = product.getQuantity();
+            } else {
+                quantity = requestDto.getQuantity();
+            }
         } else {
-            productName = requestDto.getName();
-        }
-        if (requestDto.getDescription().isEmpty()) {
-            productDescription = product.getDescription();
-        } else {
-            productDescription = requestDto.getDescription();
-        }
-        if (requestDto.getPrice() <= 0) {
-            price = product.getPrice();
-        } else {
-            price = requestDto.getPrice();
-        }
-        if (requestDto.getQuantity() <= 0) {
-            quantity = product.getQuantity();
-        } else {
-            quantity = requestDto.getQuantity();
+            return product.getSeller().getDomainName();
         }
 
         product.updateProduct(
@@ -134,7 +126,21 @@ public class SellerServiceImpl implements SellerService {
 
         productSearchRepository.save(productDocument);
 
-        return product.getSeller().getId();
+        return product.getSeller().getDomainName();
+    }
+
+    @Override
+    @Transactional
+    public String deleteProduct(Long productId, User user) {
+        Seller seller = findSeller(user);
+
+        Product product = findProduct(productId, seller);
+
+        productRepository.deleteById(product.getId());
+
+        productSearchRepository.deleteByProductId(product.getId());
+
+        return product.getSeller().getDomainName();
     }
 
     @Override
@@ -221,20 +227,6 @@ public class SellerServiceImpl implements SellerService {
         Long sellerId = sellerQuery.checkSeller(user.getId());
 
         return new SellerCheckResponse(sellerId);
-    }
-
-    @Override
-    @Transactional
-    public Long deleteProduct(Long productId, User user) {
-        Seller seller = findSeller(user);
-
-        Product product = findProduct(productId, seller);
-
-        productRepository.deleteById(product.getId());
-
-        productSearchRepository.deleteByProductId(product.getId());
-
-        return product.getSeller().getId();
     }
 
 }
